# mall-service/coupon-service/src/main/java/com/skywalker/sms/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "微服务工程代码生成器,生成Dao、Service、Controller、Pojo、Feign、Swagger的JSON脚本",
    "version": "1.0.0",
    "title": "深圳黑马训练营代码生成器",
    "termsOfService": "https://github.com/shenkunlin/code-template.git",
    "contact": {
      "email": "shenkunlin@itcast.cn"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/smsCoupon/{id}": {
      "delete": {
        "tags": [
          "SmsCouponController"
        ],
        "summary": "根据ID删除SmsCoupon",
        "description": "根据ID删除SmsCoupon方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除SmsCoupon方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除SmsCoupon"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "SmsCouponController"
        ],
        "summary": "根据ID修改SmsCoupon",
        "description": "根据ID修改SmsCoupon方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SmsCoupon方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "SmsCoupon",
            "description": "传入SmsCoupon的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsCoupon"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SmsCoupon"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "SmsCouponController"
        ],
        "summary": "根据ID查询SmsCoupon",
        "description": "根据ID查询SmsCoupon方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SmsCoupon方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SmsCoupon"
            ,
            "schema": {
              "$ref": "#/definitions/Result«SmsCoupon»"
            }
          }
        }
      }
    },
    "/smsCoupon/search": {
      "post": {
        "tags": [
          "SmsCouponController"
        ],
        "summary": "不带分页条件搜索SmsCoupon",
        "description": "不带分页条件搜索SmsCoupon方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "SmsCoupon",
            "description": "传入SmsCoupon的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsCoupon"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SmsCoupon"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsCoupon»»"
            }
          }
        }
      }
    },
    "/smsCoupon/search/{page}/{size}": {
      "get": {
        "tags": [
          "SmsCouponController"
        ],
        "summary": "分页列表查询SmsCoupon",
        "description": "分页列表查询SmsCoupon方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SmsCoupon"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsCoupon»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "SmsCouponController"
        ],
        "summary": "分页条件搜索SmsCoupon",
        "description": "分页条件搜索SmsCoupon方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "SmsCoupon",
            "description": "传入SmsCoupon的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsCoupon"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索SmsCoupon"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsCoupon»»"
            }
          }
        }
      }
    },
...
    "/smsCouponHistory/{id}": {
      "delete": {
        "tags": [
          "SmsCouponHistoryController"
        ],
        "summary": "根据ID删除SmsCouponHistory",
        "description": "根据ID删除SmsCouponHistory方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除SmsCouponHistory方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除SmsCouponHistory"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "SmsCouponHistoryController"
        ],
        "summary": "根据ID修改SmsCouponHistory",
        "description": "根据ID修改SmsCouponHistory方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SmsCouponHistory方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "SmsCouponHistory",
            "description": "传入SmsCouponHistory的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsCouponHistory"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SmsCouponHistory"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "SmsCouponHistoryController"
        ],
        "summary": "根据ID查询SmsCouponHistory",
        "description": "根据ID查询SmsCouponHistory方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SmsCouponHistory方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SmsCouponHistory"
            ,
            "schema": {
              "$ref": "#/definitions/Result«SmsCouponHistory»"
            }
          }
        }
      }
    },
    "/smsCouponHistory/search": {
      "post": {
        "tags": [
          "SmsCouponHistoryController"
        ],
        "summary": "不带分页条件搜索SmsCouponHistory",
        "description": "不带分页条件搜索SmsCouponHistory方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "SmsCouponHistory",
            "description": "传入SmsCouponHistory的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsCouponHistory"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SmsCouponHistory"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsCouponHistory»»"
            }
          }
        }
      }
    },
    "/smsCouponHistory/search/{page}/{size}": {
      "get": {
        "tags": [
          "SmsCouponHistoryController"
        ],
        "summary": "分页列表查询SmsCouponHistory",
        "description": "分页列表查询SmsCouponHistory方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SmsCouponHistory"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsCouponHistory»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "SmsCouponHistoryController"
        ],
        "summary": "分页条件搜索SmsCouponHistory",
        "description": "分页条件搜索SmsCouponHistory方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "SmsCouponHistory",
            "description": "传入SmsCouponHistory的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsCouponHistory"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索SmsCouponHistory"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsCouponHistory»»"
            }
          }
        }
      }
    },
...
    "/smsCouponSpuCategoryRelation/{id}": {
      "delete": {
        "tags": [
          "SmsCouponSpuCategoryRelationController"
        ],
        "summary": "根据ID删除SmsCouponSpuCategoryRelation",
        "description": "根据ID删除SmsCouponSpuCategoryRelation方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除SmsCouponSpuCategoryRelation方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除SmsCouponSpuCategoryRelation"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "SmsCouponSpuCategoryRelationController"
        ],
        "summary": "根据ID修改SmsCouponSpuCategoryRelation",
        "description": "根据ID修改SmsCouponSpuCategoryRelation方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SmsCouponSpuCategoryRelation方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "SmsCouponSpuCategoryRelation",
            "description": "传入SmsCouponSpuCategoryRelation的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsCouponSpuCategoryRelation"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SmsCouponSpuCategoryRelation"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "SmsCouponSpuCategoryRelationController"
        ],
        "summary": "根据ID查询SmsCouponSpuCategoryRelation",
        "description": "根据ID查询SmsCouponSpuCategoryRelation方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SmsCouponSpuCategoryRelation方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SmsCouponSpuCategoryRelation"
            ,
            "schema": {
              "$ref": "#/definitions/Result«SmsCouponSpuCategoryRelation»"
            }
          }
        }
      }
    },
    "/smsCouponSpuCategoryRelation/search": {
      "post": {
        "tags": [
          "SmsCouponSpuCategoryRelationController"
        ],
        "summary": "不带分页条件搜索SmsCouponSpuCategoryRelation",
        "description": "不带分页条件搜索SmsCouponSpuCategoryRelation方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "SmsCouponSpuCategoryRelation",
            "description": "传入SmsCouponSpuCategoryRelation的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsCouponSpuCategoryRelation"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SmsCouponSpuCategoryRelation"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsCouponSpuCategoryRelation»»"
            }
          }
        }
      }
    },
    "/smsCouponSpuCategoryRelation/search/{page}/{size}": {
      "get": {
        "tags": [
          "SmsCouponSpuCategoryRelationController"
        ],
        "summary": "分页列表查询SmsCouponSpuCategoryRelation",
        "description": "分页列表查询SmsCouponSpuCategoryRelation方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SmsCouponSpuCategoryRelation"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsCouponSpuCategoryRelation»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "SmsCouponSpuCategoryRelationController"
        ],
        "summary": "分页条件搜索SmsCouponSpuCategoryRelation",
        "description": "分页条件搜索SmsCouponSpuCategoryRelation方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "SmsCouponSpuCategoryRelation",
            "description": "传入SmsCouponSpuCategoryRelation的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsCouponSpuCategoryRelation"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索SmsCouponSpuCategoryRelation"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsCouponSpuCategoryRelation»»"
            }
          }
        }
      }
    },
...
    "/smsCouponSpuRelation/search": {
      "post": {
        "tags": [
          "SmsCouponSpuRelationController"
        ],
        "summary": "不带分页条件搜索SmsCouponSpuRelation",
        "description": "不带分页条件搜索SmsCouponSpuRelation方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "SmsCouponSpuRelation",
            "description": "传入SmsCouponSpuRelation的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsCouponSpuRelation"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SmsCouponSpuRelation"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsCouponSpuRelation»»"
            }
          }
        }
      }
    },
...
    "/smsHomeAdv/search": {
      "post": {
        "tags": [
          "SmsHomeAdvController"
        ],
        "summary": "不带分页条件搜索SmsHomeAdv",
        "description": "不带分页条件搜索SmsHomeAdv方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "SmsHomeAdv",
            "description": "传入SmsHomeAdv的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsHomeAdv"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SmsHomeAdv"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsHomeAdv»»"
            }
          }
        }
      }
    },
...
    "/smsHomeSubject/search": {
      "post": {
        "tags": [
          "SmsHomeSubjectController"
        ],
        "summary": "不带分页条件搜索SmsHomeSubject",
        "description": "不带分页条件搜索SmsHomeSubject方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "SmsHomeSubject",
            "description": "传入SmsHomeSubject的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SmsHomeSubject"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SmsHomeSubject"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SmsHomeSubject»»"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Result": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "object",
          "description": "逻辑数据"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
...
    "Result«SmsCoupon»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "description": "逻辑数据",
          "$ref": "#/definitions/SmsCoupon"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
...
    "SmsCouponSpuCategoryRelation": {
      "type": "object",
      "properties": {
        "id": {
          "type": "long",
          "description": "id"
        },
        "couponId": {
          "type": "long",
          "description": "优惠券id"
        },
        "categoryId": {
          "type": "long",
          "description": "产品分类id"
        },
        "categoryName": {
          "type": "string",
          "description": "产品分类名称"
        }
      },
      "description": "SmsCouponSpuCategoryRelation"
    },
...
  },
...
}